FAB-387 Direct search journey now work as agreed. The users search information displays with the ability to modify individual search criteria. Back button navigation had been fixed in line with agreed functionality.
If a user decides to go back to a previous answer, ensure information already entered is pre-populated.
The answers pages should look the same as they did when the user last used them.
When they’ve finished, the ‘Continue’ button should return them to the check answers page.
They should not need to go through the rest of the transaction again.